+++ server/rules/internal/catalog/platforms.go
@@ -18,6 +18,7 @@ func (*PrivilegeLaunchdPlistWrite) Platforms() []api.Platform {
 	return []api.Platform{api.PlatformDarwin}
 }
 func (*SudoersTamper) Platforms() []api.Platform           { return []api.Platform{api.PlatformDarwin} }
+func (*SudoersDestroyed) Platforms() []api.Platform        { return []api.Platform{api.PlatformDarwin} }
 func (*ApplicationControlBlock) Platforms() []api.Platform { return []api.Platform{api.PlatformDarwin} }
 func (*DNSC2Beacon) Platforms() []api.Platform             { return []api.Platform{api.PlatformDarwin} }
 func (*SensorTamper) Platforms() []api.Platform            { return []api.Platform{api.PlatformDarwin} }